+++ packages/Angular/Generic/gantt/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @memberjunction/ng-gantt
 
+## 5.30.1
+
+### Patch Changes
+
+- @memberjunction/global@5.30.1
+
+## 5.30.0
+
+### Patch Changes
+
+- @memberjunction/global@5.30.0
+
 ## 5.29.0
 
 ### Patch Changes